+++ lib_shopware6_api_base/lib_shopware6_api_base.py
@@ -512,8 +512,6 @@ class Shopware6AdminAPIClientBase(object):
         >>> my_payload=Criteria()
         >>> my_response_dict = my_api_client.request_get_paginated(request_url='product', payload=my_payload, junk_size=10)
         >>> assert 5 < len(my_response_dict['data'])
-
-
         """
         # admin_api_get_paginated}}}
         response_dict = self._request_paginated(
@@ -633,6 +631,36 @@ class Shopware6AdminAPIClientBase(object):
         >>> my_response_dict = my_api_client.request_post_paginated(request_url=my_url, payload=my_payload, junk_size=3)
         >>> assert 4 == len(my_response_dict['data'])
 
+        >>> # search for orders
+        >>> # test https://github.com/bitranox/lib_shopware6_api_base/issues/11
+        >>> import pprint
+        >>> date_from = '2024-09-29T00:00:00.000Z'
+        >>> date_to = '2024-09-29T23:59:59.999Z'
+        >>> my_criteria = Criteria()
+        >>> my_criteria.filter.append(RangeFilter(field="orderDate", parameters = {'gte': date_from, 'lte': date_to}))
+        >>> my_criteria.filter.append(MultiFilter('or', [
+        ...     EqualsFilter(field='documents.documentType.technicalName', value='invoice'),
+        ...     EqualsFilter(field='documents.documentType.technicalName', value='storno')]))
+        >>> pprint_attrs(my_criteria)
+        {'limit': None,
+         'page': None,
+         'filter': [{'type': 'range',
+                     'field': 'orderDate',
+                     'parameters': {'gte': '2024-09-29T00:00:00.000Z',
+                                    'lte': '2024-09-29T23:59:59.999Z'}},
+                    {'type': 'multi',
+                     'operator': 'or',
+                     'queries': [{'type': 'equals',
+                                  'field': 'documents.documentType.technicalName',
+                                  'value': 'invoice'},
+                                 {'type': 'equals',
+                                  'field': 'documents.documentType.technicalName',
+                                  'value': 'storno'}]}],
+         'term': None,
+         'total_count_mode': None}
+        >>> my_response_dict = my_api_client.request_post_paginated(request_url='search/order', payload=my_criteria)
+        >>> pprint.pprint(my_response_dict)
+        {'data': []}
         """
         # admin_api_post_paginated}}}
         response_dict = self._request_paginated(
@@ -711,8 +739,8 @@ class Shopware6AdminAPIClientBase(object):
         self, http_method: str, request_url: str, payload: PayLoad = None, junk_size: int = 100, update_header_fields: Optional[Dict[str, str]] = None
     ) -> Dict[str, Any]:
         """
-        request the data paginated - metadata 'total' and 'totalCountMode' will be updated
-        the paginated request reads those records in junks of junk_size=100 for performance reasons.
+        request the data paginated for performance reasons - metadata 'total' and 'totalCountMode' will be updated
+        the paginated request reads all records in junks of junk_size=100 up to "limit"
 
         payload "limit" will be respected (meaning we deliver only 'limit' results back)
         "page" will be ignored